Refactor BeamReader to use SharedPointerPool
Using SharedPointerPool<IwrfTsPulse> to implement the pulse pool
eliminates the need for explicit reference counting and moving of
pulses back for recycling. The implementation is also faster since
finding in-use nodes which are being returned to the free list is
is now a O(log N) process rather than O(N).